Factor out answer buttons component
